Create audio layer
This commit further separates the current audio code by creating a new
runtime layer and creating an object library in CMake. That additionalyl
required moving out common headers to a new library “libmilg” in order
to have clean includes, which, in turn, caused the asset store become a
collection of free functions, removing the need for hauling an object
from the application instance.