Implement typed loaders
This commit rewrites the asset system yet again to offer typed loaders
that populate the store with ready-to-use objects (e.g. textures) and
adapts existing code to the changes. As a result, the map loading is
moved out to a separate loader and asset lists for preloading are no
more.